BankAccountOOPMultiuser: DDD structure already implemented. Need to add a few more functionalities to the menu. Later on I will be adding tests and implementing a WEB API version

+++ BankAccountOOPMultiuserDDD/BankAccountOOPMultiuser.Infrastructure.Contracts/Entities/AccountEntity.cs
@@ -3,8 +3,8 @@
     public class AccountEntity
     {
         public uint Id;
-        public int Balance;
-        public List<Tuple<DateTime, int>> Movements;
+        public decimal Balance;
+        public List<Tuple<DateTime, decimal>> Movements;
         public string Pin;
         public string IBAN;
     }
